feat(web): toggleable Scheduler entry under each project in left sidebar
Adds a new project-level scheduler_view feature flag (mirroring cycle_view,
module_view, etc.). When admins enable it from Project settings → Features,
a "Scheduler" item appears in the sidebar under the project and routes to a
new top-level /<workspace>/projects/<id>/schedulers page that reuses the
existing scheduler bindings UI from Settings → Schedulers.

- Backend: add Project.scheduler_view (default False) + migration 0135;
  surface it in ProjectCreateSerializer and the project list endpoint's
  values() projection.
- Frontend: add scheduler_view to IPartialProject and EProjectFeatureKey;
  wire the new sidebar item into the three project-navigation registries
  (sidebar, ce hook, ce helper); register the new route in routes/core.ts;
  extract the bindings UI into ProjectSchedulerBindingsList so both the
  settings page and the new sidebar route render identical content.
