+++ apps/api/pi_dash/app/views/project/base.py
@@ -182,6 +182,7 @@ class ProjectViewSet(BaseViewSet):
             "module_view",
             "page_view",
             "inbox_view",
+            "scheduler_view",
             "guest_view_all_features",
             "project_lead",
             "network",
