+++ apps/web/ce/components/projects/navigation/helper.tsx
@@ -6,7 +6,15 @@
 
 // pi dash imports
 import { EUserPermissions, EProjectFeatureKey } from "@pi-dash/constants";
-import { CycleIcon, IntakeIcon, ModuleIcon, PageIcon, ViewsIcon, WorkItemsIcon } from "@pi-dash/propel/icons";
+import {
+  CalendarAfterIcon,
+  CycleIcon,
+  IntakeIcon,
+  ModuleIcon,
+  PageIcon,
+  ViewsIcon,
+  WorkItemsIcon,
+} from "@pi-dash/propel/icons";
 // components
 import type { TNavigationItem } from "@/components/workspace/sidebar/project-navigation";
 
@@ -19,6 +27,7 @@ export const getProjectFeatureNavigation = (
     issue_views_view: boolean;
     page_view: boolean;
     inbox_view: boolean;
+    scheduler_view: boolean;
   }
 ): TNavigationItem[] => [
   {
@@ -80,5 +89,15 @@ export const getProjectFeatureNavigation = (
     access: [EUserPermissions.ADMIN, EUserPermissions.MEMBER, EUserPermissions.GUEST],
     shouldRender: project.inbox_view,
     sortOrder: 6,
+  },
+  {
+    i18n_key: "sidebar.schedulers",
+    key: EProjectFeatureKey.SCHEDULERS,
+    name: "Scheduler",
+    href: `/${workspaceSlug}/projects/${projectId}/schedulers`,
+    icon: CalendarAfterIcon,
+    access: [EUserPermissions.ADMIN],
+    shouldRender: project.scheduler_view,
+    sortOrder: 7,
   },
 ];
