+++ packages/constants/src/project.ts
@@ -128,4 +128,5 @@ export enum EProjectFeatureKey {
   VIEWS = "views",
   PAGES = "pages",
   INTAKE = "intake",
+  SCHEDULERS = "schedulers",
 }
